finish lab3

+++ lab3/utils.py
@@ -21,7 +21,7 @@ def decode_sequence(ix_to_word, seq):
         txt = ''
         for j in range(D):
             ix = seq[i,j]
-            if ix > 0 :
+            if ix > 0:
                 if j >= 1:
                     txt = txt + ' '
                 txt = txt + ix_to_word[str(ix)]
@@ -41,7 +41,7 @@ class LanguageModelCriterion(nn.Module):
         target = target.contiguous().view(-1, 1)
         mask = (target != 0)
 
-        target.data = target.data.long()
+        target.data = target.data
         mask.data = mask.data.float()
         output = - input.gather(1, target) * mask
         output = torch.sum(output) / torch.sum(mask)
